refactor(cli): extract cli_common + cli_citations from the god-file (ARCH-2 M1a)
First slice of the cli.py (8.1k-LOC) decomposition. Pure mechanical move (13
functions, reviewer-verified byte-equivalent, full suite 3113 green) of
cross-cutting helpers -> cli_common.py and cite/quote/compose handlers ->
cli_citations.py. Mechanical move delegated to Codex (gpt-5.5); Claude reviewed
the diff and fixed the get_config pattern.

Re-export: cli.py re-exports _cite so `from research_hub.cli import _cite`
(mcp_server) keeps working; build_parser/_cmd_install/main remain in cli.py.

get_config sync (Option A -- the canonical pattern for M1b/M2): a single
_sync_cli_dependencies() called once at the top of _main_dispatch propagates
cli.py's (possibly test-patched) get_config into the extracted modules.
Required because the conftest autouse fixture + many tests patch
research_hub.cli.get_config while moved handlers bind their own at import.
Replaces a per-handler runtime-injection hack from the first Codex pass.

Contract tests (tests/test_arch2_cli_split.py): re-export + sync propagation.

Follow-up (fold into M1b): drop imports in cli.py orphaned by the move
(7 writing symbols + nullcontext/redirect_stdout/is_dataclass/warn_deprecated).

Co-Reviewed-By: code-reviewer skill (ARCH-2 M1a: pure-move + sync verified)
