+++ src/research_hub/cli.py
@@ -47,105 +47,41 @@ from research_hub.writing import (
     resolve_paper_meta,
     save_quote,
 )
+from research_hub import cli_citations as _cli_citations
+from research_hub.cli_common import (
+    _cli_deprecated_alias,
+    _emit_cli_json,
+    _json_safe,
+    _read_zotero_key_from_frontmatter,
+    _stdout_to_stderr,
+    _warn_cli_deprecated_alias_from_args,
+    _warn_cli_deprecated_alias_from_argv,
+)
+from research_hub.cli_citations import (
+    _cite,
+    _collect_paper_meta_for_cluster,
+    _compose_draft,
+    _quote_add,
+    _quote_list,
+    _quote_remove,
+)
 
 
-_CLI_DEPRECATED_ALIASES: dict[tuple[str, ...], tuple[str, str]] = {
-    ("ask",): ("research-hub ask", "research-hub notebooklm ask"),
-    ("summarize",): ("research-hub summarize", "research-hub paper summarize"),
-    ("cleanup",): ("research-hub cleanup", "research-hub tidy"),
-    ("label-bulk",): ("research-hub label-bulk", "research-hub paper bulk-relabel"),
-}
+def _sync_cli_dependencies() -> None:
+    """Propagate cli.py's (possibly test-patched) ``get_config`` into the
+    extracted ``cli_*`` domain modules. The conftest autouse fixture and many
+    tests patch ``research_hub.cli.get_config``; handlers that now live outside
+    cli.py bind their own ``get_config`` at import, so without this sync they
+    would not see the patch. Called ONCE at the top of ``_main_dispatch``. As
+    more modules are extracted (M1b/M2), add one line per module here."""
+    _cli_citations.get_config = get_config
+
 
 _PEER_REVIEWED_HELP = (
     "Peer-reviewed only: drop preprint backends, exclude preprint/report/dataset "
     "doc types, require >=1-backend corroboration. Excludes gray literature "
     "(arXiv/bioRxiv/Zenodo)."
 )
-
-
-def _cli_deprecated_alias(argv: list[str] | tuple[str, ...]) -> tuple[str, str] | None:
-    if not argv:
-        return None
-    command = argv[0]
-    for alias, deprecation in _CLI_DEPRECATED_ALIASES.items():
-        if (command,) == alias:
-            return deprecation
-    return None
-
-
-def _warn_cli_deprecated_alias_from_argv(argv: list[str] | tuple[str, ...]) -> None:
-    deprecation = _cli_deprecated_alias(argv)
-    if deprecation is None:
-        return
-    what, replacement = deprecation
-    warn_deprecated(
-        what,
-        replacement=replacement,
-        removed_in="v2.0.0",
-        stacklevel=3,
-    )
-
-
-def _warn_cli_deprecated_alias_from_args(args: argparse.Namespace) -> None:
-    _warn_cli_deprecated_alias_from_argv([str(args.command or "")])
-
-
-def _json_safe(value):
-    if hasattr(value, "to_dict") and callable(getattr(value, "to_dict")):
-        return _json_safe(value.to_dict())
-    if isinstance(value, Path):
-        return str(value)
-    if is_dataclass(value) and not isinstance(value, type):
-        return {key: _json_safe(item) for key, item in asdict(value).items()}
-    if isinstance(value, dict):
-        return {str(key): _json_safe(item) for key, item in value.items()}
-    if isinstance(value, (list, tuple, set)):
-        return [_json_safe(item) for item in value]
-    return value
-
-
-def _emit_cli_json(command: str, rc: int, report) -> None:
-    """Emit a structured JSON envelope for any CLI subcommand under --json.
-
-    Envelope shape (v0.91.0 W5 contract, schema_version 1.0):
-
-        {
-          "schema_version": "1.0",
-          "ok": bool,                 # rc == 0
-          "command": str,             # subcommand name, e.g. "auto"
-          "version": str,             # research_hub.__version__
-          "report": <_json_safe(report)>,   # per-command shape
-        }
-
-    The ENVELOPE is versioned and stable. The `report` payload shape
-    is per-command — agents reasoning against `_emit_cli_json` output
-    should branch on `command` to know what to expect inside `report`.
-    Per-Report schema_version is tracked for v0.92 (G2 audit #8 follow-up).
-
-    Forward-compat: pre-v0.91 envelopes lack `schema_version` entirely.
-    A reader that sees the key MISSING should treat the envelope as
-    schema 0 (legacy) and continue — the {ok, command, version, report}
-    keys are present in every version since v0.89.0.
-
-    v0.89.1: default=str catches the edge cases _json_safe returns
-    verbatim (datetime, bytes, Exception instances, custom objects)
-    — the v0.89.0 code-review skill flagged these would crash
-    json.dumps. Cheap belt-and-suspenders.
-    """
-    from research_hub import __version__
-
-    payload = {
-        "schema_version": "1.0",
-        "ok": rc == 0,
-        "command": command,
-        "version": __version__,
-        "report": _json_safe(report),
-    }
-    print(json.dumps(payload, indent=2, ensure_ascii=False, default=str))
-
-
-def _stdout_to_stderr(enabled: bool):
-    return redirect_stdout(sys.stderr) if enabled else nullcontext()
 
 
 def _config_encrypt_secrets() -> int:
@@ -2726,231 +2662,6 @@ def _synthesize(cluster: str | None, graph_colors: bool) -> int:
         print(f"Updated graph.json with {refresh_graph_from_vault(cfg)} color groups")
 
     return 0
-
-
-def _collect_paper_meta_for_cluster(cfg, cluster: str) -> list[dict]:
-    cluster_dir = cfg.raw / cluster
-    if not cluster_dir.exists():
-        raise FileNotFoundError(f"Cluster folder not found: {cluster_dir}")
-    return [format_paper_meta_from_frontmatter(path) for path in sorted(cluster_dir.glob("*.md"))]
-
-
-def _cite(
-    identifier: str | None,
-    cluster: str | None,
-    content_format: str,
-    out_path: str | None,
-    *,
-    inline: bool = False,
-    markdown: bool = False,
-    style: str = "apa",
-) -> int:
-    """Export BibTeX / BibLaTeX / RIS / CSL-JSON for a paper or cluster.
-
-    Resolves the identifier (DOI, slug, or raw title) to one or more
-    Zotero item keys via the dedup index and vault frontmatter, then
-    calls ZoteroDualClient.get_formatted to fetch each entry. Concatenates
-    results and writes to stdout or --out file.
-    """
-    from research_hub.dedup import normalize_doi
-    from research_hub.zotero.client import ZoteroDualClient
-
-    cfg = get_config()
-
-    if inline or markdown:
-        if cluster:
-            try:
-                metas = _collect_paper_meta_for_cluster(cfg, cluster)
-            except FileNotFoundError as exc:
-                print(str(exc))
-                return 1
-            rendered = []
-            for meta in metas:
-                if markdown:
-                    rendered.append(build_markdown_citation(meta))
-                else:
-                    rendered.append(build_inline_citation(meta, style=style))
-            body = "\n".join(item for item in rendered if item)
-            if not body:
-                print(f"No notes found in cluster '{cluster}'")
-                return 1
-            if out_path:
-                Path(out_path).write_text(body + "\n", encoding="utf-8")
-                print(f"Wrote {len(rendered)} citations to {out_path}")
-            else:
-                print(body)
-            return 0
-
-        if not identifier:
-            print("Either a positional <identifier> or --cluster <slug> is required")
-            return 2
-        meta = resolve_paper_meta(cfg, identifier)
-        if not meta:
-            print(f"Could not resolve identifier '{identifier}'")
-            return 1
-        body = build_markdown_citation(meta) if markdown else build_inline_citation(meta, style=style)
-        if out_path:
-            Path(out_path).write_text(body + "\n", encoding="utf-8")
-            print(f"Wrote citation to {out_path}")
-        else:
-            print(body)
-        return 0
-
-    index = DedupIndex.load(cfg.research_hub_dir / "dedup_index.json")
-
-    keys: list[str] = []
-    if cluster:
-        cluster_dir = cfg.raw / cluster
-        if not cluster_dir.exists():
-            print(f"Cluster folder not found: {cluster_dir}")
-            return 1
-        for md_path in sorted(cluster_dir.glob("*.md")):
-            key = _read_zotero_key_from_frontmatter(md_path)
-            if key:
-                keys.append(key)
-        if not keys:
-            print(f"No zotero-key entries found in {cluster_dir}")
-            return 1
-    elif identifier:
-        normalized = normalize_doi(identifier)
-        hits = index.doi_to_hits.get(normalized, [])
-        for hit in hits:
-            if hit.zotero_key and hit.zotero_key not in keys:
-                keys.append(hit.zotero_key)
-        if not keys:
-            # Fall back: treat identifier as a filename stem in raw/
-            for md_path in cfg.raw.rglob(f"{identifier}.md"):
-                key = _read_zotero_key_from_frontmatter(md_path)
-                if key:
-                    keys.append(key)
-        if not keys:
-            print(f"Could not resolve identifier '{identifier}' to a Zotero key")
-            return 1
-    else:
-        print("Either a positional <identifier> or --cluster <slug> is required")
-        return 2
-
-    dual = ZoteroDualClient()
-    entries: list[str] = []
-    for key in keys:
-        try:
-            entries.append(dual.get_formatted(key, content_format=content_format))
-        except Exception as exc:
-            print(f"  [warn] {key}: {exc}")
-    body = "\n\n".join(e for e in entries if e)
-    if out_path:
-        Path(out_path).write_text(body + "\n", encoding="utf-8")
-        print(f"Wrote {len(entries)} {content_format} entries to {out_path}")
-    else:
-        print(body)
-    return 0 if entries else 1
-
-
-def _quote_add(slug: str, page: str, text: str, context: str) -> int:
-    cfg = get_config()
-    meta = resolve_paper_meta(cfg, slug)
-    quote = Quote(
-        slug=str(meta.get("slug", slug) or slug),
-        doi=str(meta.get("doi", "") or ""),
-        title=str(meta.get("title", slug) or slug),
-        authors=str(meta.get("authors", "") or ""),
-        year=str(meta.get("year", "") or ""),
-        cluster_slug=str(meta.get("topic_cluster", "") or ""),
-        page=page,
-        text=text,
-        context_note=context,
-    )
-    path = save_quote(cfg, quote)
-    print(path)
-    return 0
-
-
-def _quote_list(cluster: str | None) -> int:
-    cfg = get_config()
-    quotes = load_all_quotes(cfg)
-    if cluster:
-        quotes = [quote for quote in quotes if quote.cluster_slug == cluster]
-    for quote in quotes:
-        text = re.sub(r"\s+", " ", quote.text).strip()
-        preview = text[:80] + ("..." if len(text) > 80 else "")
-        print(f"{quote.slug}\t{quote.captured_at}\t{quote.page}\t{preview}")
-    return 0
-
-
-def _quote_remove(slug: str, at: str) -> int:
-    cfg = get_config()
-    path = cfg.research_hub_dir / "quotes" / f"{slug}.md"
-    if not path.exists():
-        print(f"Quote file not found: {path}")
-        return 1
-    original = path.read_text(encoding="utf-8")
-    blocks = list(re.finditer(r"^---\n.*?\n---\n.*?(?:\n(?=---\n)|\Z)", original, re.DOTALL | re.MULTILINE))
-    kept: list[str] = []
-    removed = 0
-    for match in blocks:
-        block = match.group(0).strip()
-        if f"captured_at: {at}" in block and removed == 0:
-            removed += 1
-            continue
-        kept.append(block)
-    if removed == 0:
-        print(f"No quote block found for {slug} at {at}")
-        return 1
-    if kept:
-        path.write_text("\n\n".join(kept) + "\n", encoding="utf-8")
-    else:
-        path.unlink()
-    print(f"Removed quote {slug} at {at}")
-    return 0
-
-
-def _compose_draft(
-    cluster_slug: str,
-    outline: str | None,
-    quotes: str | None,
-    style: str,
-    include_bibliography: bool,
-    out: str | None,
-) -> int:
-    from research_hub.drafting import DraftingError, compose_draft_from_cli
-
-    cfg = get_config()
-    try:
-        result = compose_draft_from_cli(
-            cfg,
-            cluster_slug,
-            outline=outline,
-            quote_slugs=quotes,
-            style=style,
-            include_bibliography=include_bibliography,
-            out=out,
-        )
-    except DraftingError as exc:
-        print(f"Draft composition failed: {exc}")
-        return 1
-    print(f"Draft written to {result.path}")
-    print(
-        f"  {result.quote_count} quotes, {result.cited_paper_count} cited papers, "
-        f"{result.section_count} sections"
-    )
-    return 0
-
-
-def _read_zotero_key_from_frontmatter(md_path: Path) -> str | None:
-    """Pull the `zotero-key: XXXX` line out of an Obsidian raw note."""
-    try:
-        text = md_path.read_text(encoding="utf-8", errors="ignore")
-    except OSError:
-        return None
-    if not text.startswith("---"):
-        return None
-    end = text.find("\n---", 3)
-    if end < 0:
-        return None
-    frontmatter = text[3:end]
-    import re as _re
-    match = _re.search(r"^zotero-key:\s*([A-Z0-9]+)", frontmatter, _re.MULTILINE)
-    return match.group(1) if match else None
 
 
 def _read_doi_from_frontmatter(md_path: Path) -> str | None:
@@ -7420,6 +7131,10 @@ def _main_dispatch(args, parser) -> int:
     if args.command is None:
         parser.print_help()
         return 0
+
+    # Propagate any test-patched get_config into the extracted cli_* modules
+    # before dispatching to a handler that may live outside cli.py.
+    _sync_cli_dependencies()
 
     _warn_cli_deprecated_alias_from_args(args)
